Refactor: Splitting Select and Move utils
